requirements.txt
Django>=4.2,<5.0
psycopg2-binary
djangorestframework>=3.14,<4.0